Fix cqlsh history problem (bash hopefully as well)
Given how cqlsh tries to find the history, the only way is to set HOME
to something where the default docker user has permissions to write.

This ends up on the data volume, but should not be an issue since if you
got one, you get a lot of the actual data already.