Fix user.home property for running under Taupage
Ever Since we've fixed the HOME/current directory, nodetool tries to
save its history in a file named $HOME/.cassandra/nodetool.history

But because of the way Taupage runs docker, value of "user.home" system
property under Java is returned as "?".  Which makes it create literally
a directory named "?" in /var/lib/cassandra, which is very confusing.

It is not apparent why nodetool tries to save its history in a file in
the first place and there doesn't seem to be a flag to turn it off.