Update documentation on disabling default Cassandra superuser
Dropping user 'cassandra' doesn't seem to be a reliable advice: we've seen the
user reappearing under some circumstances.  The official documentation says,
though: remove the superuser flag from it and set the password to something
not easily guessable.  Add links to the doc from DataStax (version 2 and 3).

Don't duplicate auth setup instructions in the README: follow the on-screen
instruction printed by the create command instead.